refactor: co-locate apex_features/split_product compute into their blocks
Round-3 review follow-up. The PR's thesis — a score family is one file that
owns its whole lifecycle, with structural (not test-based) drift protection —
was violated by its biggest family: apex_features' compute lived in
scores/apex_features.rs and its projection was hand-written (the one family
where the original silent-drift bug could still bite, caught only by the
goldens). "Compute is apex-stage, so it must live elsewhere" was a false
dichotomy: the compute has a single caller (apex_finding), so relocating it
changes when nothing runs, only where the code sits.

Changes:
- ApexFeatures is now a score_block! struct (11 #[raw] fields) — hand-written
  22-line columns/features projection deleted; the family regains the macro's
  structural drift guarantee.
- compute_apex_features + its 11 exclusive helpers + SCORING_WEIGHTS +
  compute_weighted_score move into blocks/apex_features.rs.
- compute_split_product + area_uniqueness + coelution_gradient + CoelutionScratch
  + result structs move into blocks/split_product.rs; SplitProductScore lives
  there too.
- New scoring/apex_dsp.rs holds the genuinely-shared numeric primitives
  (argmax, pearson_correlation, center_normalize, estimate_fwhm,
  build_gaussian_reference, SCRIBE_FLOOR) — reusable math with no score
  semantics, used by both families.
- scores/apex_features.rs deleted; scores/ now holds only hyperscore.
- Math tests move with their code (apex_features tests, split_product tests).

Behavior-preserving: parquet schema golden + feature-name-set golden + all
relocated math tests pass unchanged. Adding an apex feature is now a single
struct-field edit + one compute helper, both in one file.
