+++ rust/timsseek/src/scoring/blocks/mod.rs
@@ -1,22 +1,21 @@
 //! Score blocks: modular, macro-generated scoring fields.
 //!
-//! A *score* belongs to a *family*; a family is one file that owns its output
-//! lifecycle — parquet projection (`columns`) and ML projection (`features`) —
-//! and, for most families, compute-from-inputs. Adding a score to an existing
-//! family touches only that file, and forgetting the ML projection is
-//! structurally impossible: the `columns`/`features` bodies are generated by
-//! walking the same field list that defines the struct (see [`score_block!`]),
-//! so a partial emission is unrepresentable.
+//! A *score* belongs to a *family*; a family is one file that owns its whole
+//! lifecycle — compute-from-inputs, parquet projection (`columns`), and ML
+//! projection (`features`). Adding a score to an existing family touches only
+//! that file, and forgetting the ML projection is structurally impossible: the
+//! `columns`/`features` bodies are generated by walking the same field list
+//! that defines the struct (see [`score_block!`]), so a partial emission is
+//! unrepresentable. The heavy apex-stage families ([`apex_features`],
+//! [`split_product`]) keep their compute here too, leaning on the shared
+//! numeric primitives in [`crate::scoring::apex_dsp`].
 //!
-//! Hand-written exceptions:
+//! Hand-written exceptions to the macro (they still emit through the same
+//! sinks, and their compute still lives in-file):
 //! - [`identity::Identity`] — irreducibly mixed-dtype.
 //! - array families [`ion_errors::IonErrors`] /
 //!   [`relative_intensities::RelativeIntensities`] and the conditional,
-//!   features-only [`sequence_counts`] block — still emit through the sinks.
-//! - [`apex_features`] — the *projection* lives here, but its large compute
-//!   ([`crate::scoring::scores::apex_features`]) stays in the apex stage,
-//!   shared with `apex_finding`. This block wraps the already-computed value;
-//!   it is the one family whose compute is not co-located.
+//!   features-only [`sequence_counts`] block.
 //!
 //! The remaining blocks — including the output-only
 //! [`result_meta::ResultMeta`] — are macro-generated by [`score_block!`].
