+++ rust/timsseek/src/scoring/scores/mod.rs
@@ -1,2 +1,1 @@
-pub mod apex_features;
 pub mod hyperscore;
